requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
flask-wtf>=1.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
psycopg2-binary>=2.9.9
pandas>=1.3.3
numpy>=1.21.2
requests>=2.26.0
python-dotenv>=0.19.0
textblob>=0.15.3
nltk>=3.6.5
scikit-learn>=1.0.0
matplotlib>=3.4.3
seaborn>=0.11.2
gunicorn>=20.1.0
openai>=1.6.0
anthropic>=0.8.0
spacy>=3.5.0
python-dateutil 
sqlalchemy
flask-caching